static link fixes + more cleanup
