+++ Code/Engine/Foundation/Threading/Implementation/Task.cpp
@@ -58,3 +58,7 @@ void ezTask::Run(ezUInt32 uiInvocation)
 
   m_iRemainingRuns.Decrement();
 }
+
+
+EZ_STATICLINK_FILE(Foundation, Foundation_Threading_Implementation_Task);
+
