+++ Code/Engine/Foundation/Threading/Implementation/TaskGroup.cpp
@@ -44,3 +44,7 @@ void ezTaskGroup::DebugCheckTaskGroup(ezTaskGroupID groupID, ezMutex& mutex)
   EZ_ASSERT_DEV(pGroup->m_iNumActiveDependencies == 0, "Invalid active dependenices");
 }
 #endif
+
+
+EZ_STATICLINK_FILE(Foundation, Foundation_Threading_Implementation_TaskGroup);
+
