+++ Code/Engine/Foundation/Threading/Implementation/TaskSystemTasks.cpp
@@ -5,8 +5,7 @@
 #include <Foundation/Threading/Implementation/TaskWorkerThread.h>
 #include <Foundation/Threading/Lock.h>
 #include <Foundation/Threading/TaskSystem.h>
-#include <Foundation/Time/Timestamp.h>
-#include <Foundation/Utilities/DGMLWriter.h>
+
 
 void ezTaskSystem::TaskHasFinished(ezTask* pTask, ezTaskGroup* pGroup)
 {
@@ -133,43 +132,6 @@ bool ezTaskSystem::ExecuteTask(ezTaskPriority::Enum FirstPriority, ezTaskPriorit
   return true;
 }
 
-void ezTaskSystem::WaitForCondition(ezDelegate<bool()> condition)
-{
-  EZ_PROFILE_SCOPE("WaitForCondition");
-
-  EZ_ASSERT_DEV(tl_TaskWorkerInfo.m_bAllowNestedTasks, "The executing task '{}' is flagged to never wait for other tasks but does so anyway. Remove the flag or remove the wait-dependency.", tl_TaskWorkerInfo.m_szTaskName);
-
-  const auto ThreadTaskType = tl_TaskWorkerInfo.m_WorkerType;
-  const bool bAllowSleep = ThreadTaskType != ezWorkerThreadType::MainThread;
-
-  while (!condition())
-  {
-    if (!HelpExecutingTasks(ezTaskGroupID()))
-    {
-      if (bAllowSleep)
-      {
-        s_BlockedWorkerThreads[ThreadTaskType].Increment();
-
-        const ezWorkerThreadType::Enum typeToWakeUp = (ThreadTaskType == ezWorkerThreadType::Unknown) ? ezWorkerThreadType::ShortTasks : ThreadTaskType;
-
-        WakeUpThreads(typeToWakeUp, 1);
-
-        while (!condition())
-        {
-          // TODO: busy loop for now
-          ezThreadUtils::YieldTimeSlice();
-        }
-
-        s_BlockedWorkerThreads[ThreadTaskType].Decrement();
-        break;
-      }
-      else
-      {
-        ezThreadUtils::YieldTimeSlice();
-      }
-    }
-  }
-}
 
 ezResult ezTaskSystem::CancelTask(ezTask* pTask, ezOnTaskRunning::Enum OnTaskRunning)
 {
@@ -246,4 +208,216 @@ bool ezTaskSystem::HelpExecutingTasks(const ezTaskGroupID& WaitingForGroup)
   return ExecuteTask(FirstPriority, LastPriority, bOnlyTasksThatNeverWait, WaitingForGroup, nullptr);
 }
 
-EZ_STATICLINK_FILE(Foundation, Foundation_Threading_Implementation_Tasks);
+
+ezTaskGroupID ezTaskSystem::StartSingleTask(ezTask* pTask, ezTaskPriority::Enum Priority, ezTaskGroupID Dependency)
+{
+  ezTaskGroupID Group = CreateTaskGroup(Priority);
+  AddTaskGroupDependency(Group, Dependency);
+  AddTaskToGroup(Group, pTask);
+  StartTaskGroup(Group);
+  return Group;
+}
+
+ezTaskGroupID ezTaskSystem::StartSingleTask(ezTask* pTask, ezTaskPriority::Enum Priority)
+{
+  ezTaskGroupID Group = CreateTaskGroup(Priority);
+  AddTaskToGroup(Group, pTask);
+  StartTaskGroup(Group);
+  return Group;
+}
+
+void ezTaskSystem::ReprioritizeFrameTasks()
+{
+  // There should usually be no 'this frame tasks' left at this time
+  // however, while we waited to enter the lock, such tasks might have appeared
+  // In this case we move them into the highest-priority 'this frame' queue, to ensure they will be executed asap
+  for (ezUInt32 i = (ezUInt32)ezTaskPriority::ThisFrame; i <= (ezUInt32)ezTaskPriority::LateThisFrame; ++i)
+  {
+    auto it = s_Tasks[i].GetIterator();
+
+    // move all 'this frame' tasks into the 'early this frame' queue
+    while (it.IsValid())
+    {
+      s_Tasks[ezTaskPriority::EarlyThisFrame].PushBack(*it);
+
+      ++it;
+    }
+
+    // remove the tasks from their current queue
+    s_Tasks[i].Clear();
+  }
+
+  for (ezUInt32 i = (ezUInt32)ezTaskPriority::EarlyNextFrame; i <= (ezUInt32)ezTaskPriority::LateNextFrame; ++i)
+  {
+    auto it = s_Tasks[i].GetIterator();
+
+    // move all 'next frame' tasks into the 'this frame' queues
+    while (it.IsValid())
+    {
+      s_Tasks[i - 3].PushBack(*it);
+
+      ++it;
+    }
+
+    // remove the tasks from their current queue
+    s_Tasks[i].Clear();
+  }
+
+  for (ezUInt32 i = (ezUInt32)ezTaskPriority::In2Frames; i <= (ezUInt32)ezTaskPriority::In9Frames; ++i)
+  {
+    auto it = s_Tasks[i].GetIterator();
+
+    // move all 'in N frames' tasks into the 'in N-1 frames' queues
+    // moves 'In2Frames' into 'LateNextFrame'
+    while (it.IsValid())
+    {
+      s_Tasks[i - 1].PushBack(*it);
+
+      ++it;
+    }
+
+    // remove the tasks from their current queue
+    s_Tasks[i].Clear();
+  }
+}
+
+void ezTaskSystem::ExecuteSomeFrameTasks(ezUInt32 uiSomeFrameTasks, double fSmoothFrameMS)
+{
+  if (uiSomeFrameTasks == 0)
+    return;
+
+  EZ_PROFILE_SCOPE("SomeFrameMainThreadTasks");
+
+  // 'SomeFrameMainThread' tasks are usually used to upload resources that have been loaded in the background
+  // they do not need to be executed right away, but the earlier, the better
+
+  // as long as the frame time is short enough, execute tasks that need to be done on the main thread
+  // on fast machines that means that these tasks are finished as soon as possible and users will see the results quickly
+
+  // if the frame time spikes, we can skip this a few times, to try to prevent further slow downs
+  // however in such instances, the 'frame time threshold' will increase and thus the chance that we skip this entirely becomes lower over
+  // time that guarantees some progress, even if the frame rate is constantly low
+
+  static double s_fFrameTimeThreshold = fSmoothFrameMS;
+
+  static ezTime s_LastFrame; // initializes to zero -> very large frame time difference at first
+
+  ezTime CurTime = ezTime::Now();
+  ezTime LastTime = s_LastFrame;
+  s_LastFrame = CurTime;
+
+  // as long as we have a smooth frame rate, execute as many of these tasks, as possible
+  while ((uiSomeFrameTasks > 0) && ((CurTime - LastTime).GetMilliseconds() < fSmoothFrameMS))
+  {
+    // we execute one of these tasks, so reset the frame time threshold
+    s_fFrameTimeThreshold = fSmoothFrameMS;
+
+    if (!ExecuteTask(ezTaskPriority::SomeFrameMainThread, ezTaskPriority::SomeFrameMainThread, false, ezTaskGroupID(), nullptr))
+      return; // nothing left to do
+
+    CurTime = ezTime::Now();
+    --uiSomeFrameTasks;
+  }
+
+  // nothing left to do
+  if (uiSomeFrameTasks == 0)
+    return;
+
+  if ((CurTime - LastTime).GetMilliseconds() < s_fFrameTimeThreshold)
+  {
+    // we execute one of these tasks, so reset the frame time threshold
+    s_fFrameTimeThreshold = fSmoothFrameMS;
+
+    ExecuteTask(ezTaskPriority::SomeFrameMainThread, ezTaskPriority::SomeFrameMainThread, false, ezTaskGroupID(), nullptr);
+  }
+  else
+  {
+    // increase the threshold by 5 ms
+    // this means that when the frame rate is too low, we can ignore these tasks for a few frames
+    // and thus prevent decreasing the frame rate even further
+    // however we increase the time threshold, at which we skip this, further and further
+    // therefore at some point we will execute at least one such task, no matter how low the frame rate is
+    // this guarantees at least some progress with these tasks
+    s_fFrameTimeThreshold += 5.0;
+
+    //  25 ms -> 40 FPS
+    //  30 ms -> 33 FPS
+    //  35 ms -> 28 FPS
+    //  40 ms -> 25 FPS
+    //  45 ms -> 22 FPS
+    //  50 ms -> 20 FPS
+    //  55 ms -> 18 FPS
+    //  60 ms -> 16 FPS
+    //  65 ms -> 15 FPS
+    //  70 ms -> 14 FPS
+    //  75 ms -> 13 FPS
+  }
+}
+
+
+void ezTaskSystem::FinishFrameTasks()
+{
+  EZ_ASSERT_DEV(ezThreadUtils::IsMainThread(), "This function must be executed on the main thread.");
+
+  // make sure all 'main thread' and 'short' tasks are either finished or being worked on by other threads already
+  {
+    while (true)
+    {
+      // Prefer to work on main-thread tasks
+      if (ExecuteTask(ezTaskPriority::ThisFrameMainThread, ezTaskPriority::ThisFrameMainThread, false, ezTaskGroupID(), nullptr))
+      {
+        continue;
+      }
+
+      // if there are none, help out with the other tasks for this frame
+      if (ExecuteTask(ezTaskPriority::EarlyThisFrame, ezTaskPriority::LateThisFrame, false, ezTaskGroupID(), nullptr))
+      {
+        continue;
+      }
+
+      break;
+    }
+  }
+
+  ezUInt32 uiSomeFrameTasks = 0;
+
+  // all the important tasks for this frame should be finished or worked on by now
+  // so we can now re-prioritize the tasks for the next frame
+  {
+    EZ_LOCK(s_TaskSystemMutex);
+
+    // get this info once, it won't shrink (but might grow) while we are outside the lock
+    uiSomeFrameTasks = s_Tasks[ezTaskPriority::SomeFrameMainThread].GetCount();
+
+    ReprioritizeFrameTasks();
+  }
+
+  ExecuteSomeFrameTasks(uiSomeFrameTasks, s_fSmoothFrameMS);
+
+  // Update the thread utilization
+  {
+    const ezTime tNow = ezTime::Now();
+    static ezTime s_LastFrameUpdate = tNow;
+    const ezTime tDiff = tNow - s_LastFrameUpdate;
+
+    // prevent division by zero (inside ComputeThreadUtilization)
+    if (tDiff > ezTime::Seconds(0.0))
+    {
+      s_LastFrameUpdate = tNow;
+
+      for (ezUInt32 type = 0; type < ezWorkerThreadType::ENUM_COUNT; ++type)
+      {
+        const ezUInt32 uiNumWorkers = s_iNumWorkerThreads[type];
+
+        for (ezUInt32 t = 0; t < uiNumWorkers; ++t)
+        {
+          s_WorkerThreads[type][t]->UpdateThreadUtilization(tDiff);
+        }
+      }
+    }
+  }
+}
+
+
+EZ_STATICLINK_FILE(Foundation, Foundation_Threading_Implementation_TaskSystemTasks);
+
