+++ Code/Engine/Foundation/Threading/Implementation/TaskWorkerThread.cpp
@@ -127,3 +127,7 @@ double ezTaskWorkerThread::GetThreadUtilization(ezUInt32* pNumTasksExecuted /*= 
 
   return m_fLastThreadUtilization;
 }
+
+
+EZ_STATICLINK_FILE(Foundation, Foundation_Threading_Implementation_TaskWorkerThread);
+
